refactor(ui): extract date range filter into modular architecture
split the monolithic date range filter component into separate controller,
helper, and view files to improve maintainability and testability.
update imports across multiple components to reference the new modular
structure under the date_range_filter subdirectory.
